+++ app/src/main/kotlin/it/vfsfitvnm/vimusic/ui/items/ArtistItem.kt
@@ -1,0 +1,145 @@
+package it.vfsfitvnm.vimusic.ui.items
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.requiredSize
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.foundation.text.BasicText
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.text.style.TextOverflow
+import androidx.compose.ui.unit.Dp
+import androidx.compose.ui.unit.dp
+import coil.compose.AsyncImage
+import it.vfsfitvnm.vimusic.models.Artist
+import it.vfsfitvnm.vimusic.ui.components.themed.TextPlaceholder
+import it.vfsfitvnm.vimusic.ui.styling.LocalAppearance
+import it.vfsfitvnm.vimusic.ui.styling.shimmer
+import it.vfsfitvnm.vimusic.utils.secondary
+import it.vfsfitvnm.vimusic.utils.semiBold
+import it.vfsfitvnm.vimusic.utils.thumbnail
+import it.vfsfitvnm.youtubemusic.Innertube
+
+@Composable
+fun ArtistItem(
+    artist: Artist,
+    thumbnailSizePx: Int,
+    thumbnailSizeDp: Dp,
+    modifier: Modifier = Modifier,
+    alternative: Boolean = false,
+) {
+    ArtistItem(
+        thumbnailUrl = artist.thumbnailUrl,
+        name = artist.name,
+        subscribersCount = null,
+        thumbnailSizePx = thumbnailSizePx,
+        thumbnailSizeDp = thumbnailSizeDp,
+        modifier = modifier,
+        alternative = alternative
+    )
+}
+
+@Composable
+fun ArtistItem(
+    artist: Innertube.ArtistItem,
+    thumbnailSizePx: Int,
+    thumbnailSizeDp: Dp,
+    modifier: Modifier = Modifier,
+    alternative: Boolean = false,
+) {
+    ArtistItem(
+        thumbnailUrl = artist.thumbnail?.url,
+        name = artist.info?.name,
+        subscribersCount = artist.subscribersCountText,
+        thumbnailSizePx = thumbnailSizePx,
+        thumbnailSizeDp = thumbnailSizeDp,
+        modifier = modifier,
+        alternative = alternative
+    )
+}
+
+@Composable
+fun ArtistItem(
+    thumbnailUrl: String?,
+    name: String?,
+    subscribersCount: String?,
+    thumbnailSizePx: Int,
+    thumbnailSizeDp: Dp,
+    modifier: Modifier = Modifier,
+    alternative: Boolean = false,
+) {
+    val (_, typography) = LocalAppearance.current
+
+    ItemContainer(
+        alternative = alternative,
+        thumbnailSizeDp = thumbnailSizeDp,
+        horizontalAlignment = Alignment.CenterHorizontally,
+        modifier = modifier
+    ) {
+        AsyncImage(
+            model = thumbnailUrl?.thumbnail(thumbnailSizePx),
+            contentDescription = null,
+            modifier = Modifier
+                .clip(CircleShape)
+                .requiredSize(thumbnailSizeDp)
+        )
+
+        ItemInfoContainer(
+            horizontalAlignment = if (alternative) Alignment.CenterHorizontally else Alignment.Start,
+        ) {
+            BasicText(
+                text = name ?: "",
+                style = typography.xs.semiBold,
+                maxLines = if (alternative) 1 else 2,
+                overflow = TextOverflow.Ellipsis
+            )
+
+            subscribersCount?.let {
+                BasicText(
+                    text = subscribersCount,
+                    style = typography.xxs.semiBold.secondary,
+                    maxLines = 1,
+                    overflow = TextOverflow.Ellipsis,
+                    modifier = Modifier
+                        .padding(top = 4.dp)
+                )
+            }
+        }
+    }
+}
+
+@Composable
+fun ArtistItemPlaceholder(
+    thumbnailSizeDp: Dp,
+    modifier: Modifier = Modifier,
+    alternative: Boolean = false,
+) {
+    val (colorPalette) = LocalAppearance.current
+
+    ItemContainer(
+        alternative = alternative,
+        thumbnailSizeDp = thumbnailSizeDp,
+        horizontalAlignment = Alignment.CenterHorizontally,
+        modifier = modifier
+    ) {
+        Spacer(
+            modifier = Modifier
+                .background(color = colorPalette.shimmer, shape = CircleShape)
+                .size(thumbnailSizeDp)
+        )
+
+        ItemInfoContainer(
+            horizontalAlignment = if (alternative) Alignment.CenterHorizontally else Alignment.Start,
+        ) {
+            TextPlaceholder()
+            TextPlaceholder(
+                modifier = Modifier
+                    .padding(top = 4.dp)
+            )
+        }
+    }
+}
