+++ app/src/main/kotlin/it/vfsfitvnm/vimusic/ui/items/PlaylistItem.kt
@@ -1,0 +1,274 @@
+package it.vfsfitvnm.vimusic.ui.items
+
+import androidx.annotation.DrawableRes
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.BoxScope
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.requiredSize
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.foundation.text.BasicText
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.collectAsState
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.remember
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.ColorFilter
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.text.style.TextOverflow
+import androidx.compose.ui.unit.Dp
+import androidx.compose.ui.unit.dp
+import coil.compose.AsyncImage
+import it.vfsfitvnm.vimusic.Database
+import it.vfsfitvnm.vimusic.models.PlaylistPreview
+import it.vfsfitvnm.vimusic.ui.components.themed.TextPlaceholder
+import it.vfsfitvnm.vimusic.ui.styling.LocalAppearance
+import it.vfsfitvnm.vimusic.ui.styling.onOverlay
+import it.vfsfitvnm.vimusic.ui.styling.overlay
+import it.vfsfitvnm.vimusic.ui.styling.shimmer
+import it.vfsfitvnm.vimusic.utils.color
+import it.vfsfitvnm.vimusic.utils.medium
+import it.vfsfitvnm.vimusic.utils.secondary
+import it.vfsfitvnm.vimusic.utils.semiBold
+import it.vfsfitvnm.vimusic.utils.thumbnail
+import it.vfsfitvnm.youtubemusic.Innertube
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.flow.distinctUntilChanged
+import kotlinx.coroutines.flow.map
+
+@Composable
+fun PlaylistItem(
+    @DrawableRes icon: Int,
+    colorTint: Color,
+    name: String?,
+    songCount: Int?,
+    thumbnailSizeDp: Dp,
+    modifier: Modifier = Modifier,
+    alternative: Boolean = false,
+) {
+    PlaylistItem(
+        thumbnailContent = {
+            Image(
+                painter = painterResource(icon),
+                contentDescription = null,
+                colorFilter = ColorFilter.tint(colorTint),
+                modifier = Modifier
+                    .align(Alignment.Center)
+                    .size(24.dp)
+            )
+        },
+        songCount = songCount,
+        name = name,
+        channelName = null,
+        thumbnailSizeDp = thumbnailSizeDp,
+        modifier = modifier,
+        alternative = alternative
+    )
+}
+
+@Composable
+fun PlaylistItem(
+    playlist: PlaylistPreview,
+    thumbnailSizePx: Int,
+    thumbnailSizeDp: Dp,
+    modifier: Modifier = Modifier,
+    alternative: Boolean = false,
+) {
+    val thumbnails by remember {
+        Database.playlistThumbnailUrls(playlist.playlist.id).distinctUntilChanged().map {
+            it.map { url ->
+                url.thumbnail(thumbnailSizePx / 2)
+            }
+        }
+    }.collectAsState(initial = emptyList(), context = Dispatchers.IO)
+
+    PlaylistItem(
+        thumbnailContent = {
+            if (thumbnails.toSet().size == 1) {
+                AsyncImage(
+                    model = thumbnails.first().thumbnail(thumbnailSizePx),
+                    contentDescription = null,
+                    contentScale = ContentScale.Crop,
+                    modifier = it
+                )
+            } else {
+                Box(
+                    modifier = it
+                        .fillMaxSize()
+                ) {
+                    listOf(
+                        Alignment.TopStart,
+                        Alignment.TopEnd,
+                        Alignment.BottomStart,
+                        Alignment.BottomEnd
+                    ).forEachIndexed { index, alignment ->
+                        AsyncImage(
+                            model = thumbnails.getOrNull(index),
+                            contentDescription = null,
+                            contentScale = ContentScale.Crop,
+                            modifier = Modifier
+                                .align(alignment)
+                                .size(thumbnailSizeDp / 2)
+                        )
+                    }
+                }
+            }
+        },
+        songCount = playlist.songCount,
+        name = playlist.playlist.name,
+        channelName = null,
+        thumbnailSizeDp = thumbnailSizeDp,
+        modifier = modifier,
+        alternative = alternative
+    )
+}
+
+@Composable
+fun PlaylistItem(
+    playlist: Innertube.PlaylistItem,
+    thumbnailSizePx: Int,
+    thumbnailSizeDp: Dp,
+    modifier: Modifier = Modifier,
+    alternative: Boolean = false,
+) {
+    PlaylistItem(
+        thumbnailUrl = playlist.thumbnail?.url,
+        songCount = playlist.songCount,
+        name = playlist.info?.name,
+        channelName = playlist.channel?.name,
+        thumbnailSizePx = thumbnailSizePx,
+        thumbnailSizeDp = thumbnailSizeDp,
+        modifier = modifier,
+        alternative = alternative
+    )
+}
+
+@Composable
+fun PlaylistItem(
+    thumbnailUrl: String?,
+    songCount: Int?,
+    name: String?,
+    channelName: String?,
+    thumbnailSizePx: Int,
+    thumbnailSizeDp: Dp,
+    modifier: Modifier = Modifier,
+    alternative: Boolean = false,
+) {
+    PlaylistItem(
+        thumbnailContent = {
+            AsyncImage(
+                model = thumbnailUrl?.thumbnail(thumbnailSizePx),
+                contentDescription = null,
+                contentScale = ContentScale.Crop,
+                modifier = it
+            )
+        },
+        songCount = songCount,
+        name = name,
+        channelName = channelName,
+        thumbnailSizeDp = thumbnailSizeDp,
+        modifier = modifier,
+        alternative = alternative,
+    )
+}
+
+@Composable
+fun PlaylistItem(
+    thumbnailContent: @Composable BoxScope.(modifier: Modifier) -> Unit,
+    songCount: Int?,
+    name: String?,
+    channelName: String?,
+    thumbnailSizeDp: Dp,
+    modifier: Modifier = Modifier,
+    alternative: Boolean = false,
+) {
+    val (colorPalette, typography, thumbnailShape) = LocalAppearance.current
+
+    ItemContainer(
+        alternative = alternative,
+        thumbnailSizeDp = thumbnailSizeDp,
+        modifier = modifier
+    ) { centeredModifier ->
+        Box(
+            modifier = centeredModifier
+                .clip(thumbnailShape)
+                .background(color = colorPalette.background1)
+                .requiredSize(thumbnailSizeDp)
+        ) {
+            thumbnailContent(
+                modifier = Modifier
+                    .fillMaxSize()
+            )
+
+            songCount?.let {
+                BasicText(
+                    text = "$songCount",
+                    style = typography.xxs.medium.color(colorPalette.onOverlay),
+                    maxLines = 1,
+                    overflow = TextOverflow.Ellipsis,
+                    modifier = Modifier
+                        .padding(all = 4.dp)
+                        .background(color = colorPalette.overlay, shape = RoundedCornerShape(2.dp))
+                        .padding(horizontal = 4.dp, vertical = 2.dp)
+                        .align(Alignment.BottomEnd)
+                )
+            }
+        }
+
+        ItemInfoContainer(
+            horizontalAlignment = if (alternative && channelName == null) Alignment.CenterHorizontally else Alignment.Start,
+        ) {
+            BasicText(
+                text = name ?: "",
+                style = typography.xs.semiBold,
+                maxLines = 2,
+                overflow = TextOverflow.Ellipsis,
+            )
+
+            channelName?.let {
+                BasicText(
+                    text = channelName,
+                    style = typography.xs.semiBold.secondary,
+                    maxLines = 2,
+                    overflow = TextOverflow.Ellipsis,
+                )
+            }
+        }
+    }
+}
+
+@Composable
+fun PlaylistItemPlaceholder(
+    thumbnailSizeDp: Dp,
+    modifier: Modifier = Modifier,
+    alternative: Boolean = false,
+) {
+    val (colorPalette, _, thumbnailShape) = LocalAppearance.current
+
+    ItemContainer(
+        alternative = alternative,
+        thumbnailSizeDp = thumbnailSizeDp,
+        modifier = modifier
+    ) {
+        Spacer(
+            modifier = Modifier
+                .background(color = colorPalette.shimmer, shape = thumbnailShape)
+                .size(thumbnailSizeDp)
+        )
+
+        ItemInfoContainer(
+            horizontalAlignment = if (alternative) Alignment.CenterHorizontally else Alignment.Start,
+        ) {
+            TextPlaceholder()
+            TextPlaceholder()
+        }
+    }
+}
