+++ app/src/main/kotlin/it/vfsfitvnm/vimusic/ui/items/SongItem.kt
@@ -1,0 +1,204 @@
+package it.vfsfitvnm.vimusic.ui.items
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.BoxScope
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.text.BasicText
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.text.style.TextOverflow
+import androidx.compose.ui.unit.Dp
+import androidx.compose.ui.unit.dp
+import androidx.media3.common.MediaItem
+import coil.compose.AsyncImage
+import it.vfsfitvnm.vimusic.models.DetailedSong
+import it.vfsfitvnm.vimusic.ui.components.themed.TextPlaceholder
+import it.vfsfitvnm.vimusic.ui.styling.LocalAppearance
+import it.vfsfitvnm.vimusic.ui.styling.shimmer
+import it.vfsfitvnm.vimusic.utils.medium
+import it.vfsfitvnm.vimusic.utils.secondary
+import it.vfsfitvnm.vimusic.utils.semiBold
+import it.vfsfitvnm.vimusic.utils.thumbnail
+import it.vfsfitvnm.youtubemusic.Innertube
+
+@Composable
+fun SongItem(
+    song: Innertube.SongItem,
+    thumbnailSizePx: Int,
+    thumbnailSizeDp: Dp,
+    modifier: Modifier = Modifier
+) {
+    SongItem(
+        thumbnailUrl = song.thumbnail?.size(thumbnailSizePx),
+        title = song.info?.name,
+        authors = song.authors?.joinToString("") { it.name ?: "" },
+        duration = song.durationText,
+        thumbnailSizeDp = thumbnailSizeDp,
+        modifier = modifier,
+    )
+}
+
+@Composable
+fun SongItem(
+    song: MediaItem,
+    thumbnailSizeDp: Dp,
+    thumbnailSizePx: Int,
+    modifier: Modifier = Modifier,
+    onThumbnailContent: (@Composable BoxScope.() -> Unit)? = null,
+    trailingContent: (@Composable () -> Unit)? = null
+) {
+    SongItem(
+        thumbnailUrl = song.mediaMetadata.artworkUri.thumbnail(thumbnailSizePx)?.toString(),
+        title = song.mediaMetadata.title.toString(),
+        authors = song.mediaMetadata.artist.toString(),
+        duration = song.mediaMetadata.extras?.getString("durationText"),
+        thumbnailSizeDp = thumbnailSizeDp,
+        onThumbnailContent = onThumbnailContent,
+        trailingContent = trailingContent,
+        modifier = modifier,
+    )
+}
+
+@Composable
+fun SongItem(
+    song: DetailedSong,
+    thumbnailSizePx: Int,
+    thumbnailSizeDp: Dp,
+    modifier: Modifier = Modifier,
+    onThumbnailContent: (@Composable BoxScope.() -> Unit)? = null,
+    trailingContent: (@Composable () -> Unit)? = null
+) {
+    SongItem(
+        thumbnailUrl = song.thumbnailUrl?.thumbnail(thumbnailSizePx),
+        title = song.title,
+        authors = song.artistsText,
+        duration = song.durationText,
+        thumbnailSizeDp = thumbnailSizeDp,
+        onThumbnailContent = onThumbnailContent,
+        trailingContent = trailingContent,
+        modifier = modifier,
+    )
+}
+
+@Composable
+fun SongItem(
+    thumbnailUrl: String?,
+    title: String?,
+    authors: String?,
+    duration: String?,
+    thumbnailSizeDp: Dp,
+    modifier: Modifier = Modifier,
+    onThumbnailContent: (@Composable BoxScope.() -> Unit)? = null,
+    trailingContent: (@Composable () -> Unit)? = null
+) {
+    SongItem(
+        title = title,
+        authors = authors,
+        duration = duration,
+        thumbnailSizeDp = thumbnailSizeDp,
+        thumbnailContent = {
+            AsyncImage(
+                model = thumbnailUrl,
+                contentDescription = null,
+                contentScale = ContentScale.Crop,
+                modifier = Modifier
+                    .clip(LocalAppearance.current.thumbnailShape)
+                    .fillMaxSize()
+            )
+
+            onThumbnailContent?.invoke(this)
+        },
+        modifier = modifier,
+        trailingContent = trailingContent
+    )
+}
+
+@Composable
+fun SongItem(
+    thumbnailContent: @Composable BoxScope.() -> Unit,
+    title: String?,
+    authors: String?,
+    duration: String?,
+    thumbnailSizeDp: Dp,
+    modifier: Modifier = Modifier,
+    trailingContent: @Composable (() -> Unit)? = null,
+) {
+    val (_, typography) = LocalAppearance.current
+
+    ItemContainer(
+        alternative = false,
+        thumbnailSizeDp = thumbnailSizeDp,
+        modifier = modifier
+    ) {
+        Box(
+            modifier = Modifier
+                .size(thumbnailSizeDp)
+        ) {
+            thumbnailContent()
+        }
+
+        ItemInfoContainer {
+            BasicText(
+                text = title ?: "",
+                style = typography.xs.semiBold,
+                maxLines = 1,
+                overflow = TextOverflow.Ellipsis,
+            )
+
+            Row(verticalAlignment = Alignment.CenterVertically) {
+                BasicText(
+                    text = authors ?: "",
+                    style = typography.xs.semiBold.secondary,
+                    maxLines = 1,
+                    overflow = TextOverflow.Clip,
+                    modifier = Modifier
+                        .weight(1f)
+                )
+
+                duration?.let {
+                    BasicText(
+                        text = duration,
+                        style = typography.xxs.secondary.medium,
+                        maxLines = 1,
+                        overflow = TextOverflow.Ellipsis,
+                        modifier = Modifier
+                            .padding(top = 4.dp)
+                    )
+                }
+            }
+        }
+    }
+}
+
+@Composable
+fun SongItemPlaceholder(
+    thumbnailSizeDp: Dp,
+    modifier: Modifier = Modifier
+) {
+    val (colorPalette, _, thumbnailShape) = LocalAppearance.current
+
+    ItemContainer(
+        alternative = false,
+        thumbnailSizeDp =thumbnailSizeDp,
+        modifier = modifier
+    ) {
+        Spacer(
+            modifier = Modifier
+                .background(color = colorPalette.shimmer, shape = thumbnailShape)
+                .size(thumbnailSizeDp)
+        )
+
+        ItemInfoContainer {
+            TextPlaceholder()
+            TextPlaceholder()
+        }
+    }
+}
