+++ app/src/main/kotlin/it/vfsfitvnm/vimusic/ui/items/VideoItem.kt
@@ -1,0 +1,149 @@
+package it.vfsfitvnm.vimusic.ui.items
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.foundation.text.BasicText
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.text.style.TextOverflow
+import androidx.compose.ui.unit.Dp
+import androidx.compose.ui.unit.dp
+import coil.compose.AsyncImage
+import it.vfsfitvnm.vimusic.ui.components.themed.TextPlaceholder
+import it.vfsfitvnm.vimusic.ui.styling.LocalAppearance
+import it.vfsfitvnm.vimusic.ui.styling.onOverlay
+import it.vfsfitvnm.vimusic.ui.styling.overlay
+import it.vfsfitvnm.vimusic.ui.styling.shimmer
+import it.vfsfitvnm.vimusic.utils.color
+import it.vfsfitvnm.vimusic.utils.medium
+import it.vfsfitvnm.vimusic.utils.secondary
+import it.vfsfitvnm.vimusic.utils.semiBold
+import it.vfsfitvnm.youtubemusic.Innertube
+
+@Composable
+fun VideoItem(
+    video: Innertube.VideoItem,
+    thumbnailHeightDp: Dp,
+    thumbnailWidthDp: Dp,
+    modifier: Modifier = Modifier
+) {
+    VideoItem(
+        thumbnailUrl = video.thumbnail?.url,
+        duration = video.durationText,
+        title = video.info?.name,
+        uploader = video.authors?.joinToString("") { it.name ?: "" },
+        views = video.viewsText,
+        thumbnailHeightDp = thumbnailHeightDp,
+        thumbnailWidthDp = thumbnailWidthDp,
+        modifier = modifier
+    )
+}
+
+@Composable
+fun VideoItem(
+    thumbnailUrl: String?,
+    duration: String?,
+    title: String?,
+    uploader: String?,
+    views: String?,
+    thumbnailHeightDp: Dp,
+    thumbnailWidthDp: Dp,
+    modifier: Modifier = Modifier
+) {
+    val (colorPalette, typography, thumbnailShape) = LocalAppearance.current
+
+    ItemContainer(
+        alternative = false,
+        thumbnailSizeDp = 0.dp,
+        modifier = modifier
+    ) {
+        Box {
+            AsyncImage(
+                model = thumbnailUrl,
+                contentDescription = null,
+                contentScale = ContentScale.Crop,
+                modifier = Modifier
+                    .clip(thumbnailShape)
+                    .size(width = thumbnailWidthDp, height = thumbnailHeightDp)
+            )
+
+            duration?.let {
+                BasicText(
+                    text = duration,
+                    style = typography.xxs.medium.color(colorPalette.onOverlay),
+                    maxLines = 1,
+                    overflow = TextOverflow.Ellipsis,
+                    modifier = Modifier
+                        .padding(all = 4.dp)
+                        .background(color = colorPalette.overlay, shape = RoundedCornerShape(2.dp))
+                        .padding(horizontal = 4.dp, vertical = 2.dp)
+                        .align(Alignment.BottomEnd)
+                )
+            }
+        }
+
+        ItemInfoContainer {
+            BasicText(
+                text = title ?: "",
+                style = typography.xs.semiBold,
+                maxLines = 2,
+                overflow = TextOverflow.Ellipsis,
+            )
+
+            BasicText(
+                text = uploader ?: "",
+                style = typography.xs.semiBold.secondary,
+                maxLines = 1,
+                overflow = TextOverflow.Ellipsis,
+            )
+
+            views?.let {
+                BasicText(
+                    text = views,
+                    style = typography.xxs.medium.secondary,
+                    maxLines = 1,
+                    overflow = TextOverflow.Ellipsis,
+                    modifier = Modifier
+                        .padding(top = 4.dp)
+                )
+            }
+        }
+    }
+}
+
+@Composable
+fun VideoItemPlaceholder(
+    thumbnailHeightDp: Dp,
+    thumbnailWidthDp: Dp,
+    modifier: Modifier = Modifier
+) {
+    val (colorPalette, _, thumbnailShape) = LocalAppearance.current
+
+    ItemContainer(
+        alternative = false,
+        thumbnailSizeDp = 0.dp,
+        modifier = modifier
+    ) {
+        Spacer(
+            modifier = Modifier
+                .background(color = colorPalette.shimmer, shape = thumbnailShape)
+                .size(width = thumbnailWidthDp, height = thumbnailHeightDp)
+        )
+
+        ItemInfoContainer {
+            TextPlaceholder()
+            TextPlaceholder()
+            TextPlaceholder(
+                modifier = Modifier
+                    .padding(top = 8.dp)
+            )
+        }
+    }
+}
